requirements.txt
streamlit
pandas
requests
rdflib
databricks-sql-connector